refactor: extract contract governance capability normalization
