+++ docs/engineering_convergence/split_plan_queue.md
@@ -14,8 +14,8 @@ Generated from `complexity_budget_report.md` split-plan-required files.
 | Priority | Lines | Owner | File | Decomposition Direction |
 | --- | ---: | --- | --- | --- |
 | P0 | 5939 | Frontend owner | `frontend/apps/web/src/pages/ContractFormPage.vue` | Extract composables, child panels, data adapters, and action handlers; keep the route component as orchestration shell. |
-| P1 | 4535 | Platform owner | `addons/smart_core/utils/contract_governance.py` | Extract constants/registries, user-surface normalization, list governance, native bridge, form policy, diagnostics, and keep `apply_contract_governance` as a thin facade. |
 | P1 | 4372 | Construction backend owner | `addons/smart_construction_core/core_extension.py` | Define owner-specific decomposition plan before adding unrelated behavior. |
+| P1 | 4272 | Platform owner | `addons/smart_core/utils/contract_governance.py` | Extract constants/registries, user-surface normalization, list governance, native bridge, form policy, diagnostics, and keep `apply_contract_governance` as a thin facade. |
 | P1 | 3849 | Platform owner | `addons/smart_core/handlers/ui_contract_v2.py` | Extract parsing, validation, assembly, and response mapping into owned services. |
 | P1 | 3773 | Frontend owner | `frontend/apps/web/src/views/ActionView.vue` | Extract composables, child panels, data adapters, and action handlers; keep the route component as orchestration shell. |
 | P1 | 3684 | Frontend owner | `frontend/apps/web/src/views/MenuConfigView.vue` | Extract composables, child panels, data adapters, and action handlers; keep the route component as orchestration shell. |
